Share cats instances and test for Product
