+++ core/src/test/scala/io/chrisdavenport/monoids/ProductTests.scala
@@ -1,12 +1,12 @@
 package io.chrisdavenport.monoids
 
 import cats._
-import cats.implicits._
+import cats.syntax.all._
 import cats.kernel.laws.discipline._
 import cats.laws.discipline._
 
 class ProductTests
-    extends munit.DisciplineSuite
+  extends munit.DisciplineSuite
     with MonoidsArbitraries
     with ExtraMonoidsArbitraries {
   checkAll("Product", EqTests[Product[Int]].eqv)
